requirements.txt
requests~=2.31.0
pandas~=2.2.2
python-dotenv~=1.0.1
getenv~=0.2.0